Amendments made for DeleteHandler, MessageList and created a DeleteHandlerTest for testing of DeleteHandler class.
For DeleteHandler class from String array changes to Array List.
Added three if loops to check if the array list is null and empty.
Also check if the size of the array list is more than 1 or if it is an integer.

For MessageList class, changes made in variable name